testes

+++ tcs/src/main/java/go/party/tcs/repository/NotificationRepository.java
@@ -9,6 +9,6 @@ import go.party.tcs.model.Notification;
 
 @Repository
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
-    List<Notification> findByUserId(Long userId);
+    List<Notification> findByUserId(Integer userId);
 
 }
